Fixed false positive when checking unsaved, verbose commits.
When comitting code with the `--verbose` option (i.e. `git commit
--verbose`), Git will calculate the file diff and include it as part of
the commit message. The problem is that this gem would end up scanning
the contents of the included commit diff yielding false positives. This
fixes the situation so all content including (and past) the special Git
*scissors* comment is ignored for all cops.

This only affects unsaved commits as Git, by default, never includes
this information in saved commits.