Refactored affected commit body lines to abstract class.
This has become a common enough pattern that it can be moved to the
superclass in order to reduce duplicated effort within the subclasses.

When using `#affected_commit_body_lines`, the corresponding
`#invalid_line?` method *must* be implemented as well. In most cases,
subclasses will only be required to implement the `#invalid_line?`
method.

This is also a short term fix as the superclass is starting to get
overloaded due to mixed concerns. For instance, the superclass
shouldn't need to know how to deal with commit body lines -- There
should be another object that encapsulates this knowledge. Will need to
improve upon this more in the future.