Added unsaved commit trailers.
Necessary to access and process trailers for a commit. Will be used in
future style cops that will help ensure trailers use best practices.

Trailers can only be detected from the `.git/COMMIT_EDITMSG` file, for
example, when separated by a new line. Without the new line, trailers
are assumed to be part of the body which is why the fixture was updated
to mimic expected behavior by the `git-interpret-trailers` command.