Fixed multpile line commit messages for specs.
Previously, the use of `$` needed to prefix a commit message
(especially one with multiple lines) in order for the commit message to
be created properly for testing purposes. This no longer appears to be
the case (especially when testing on Circle CI).

The use of `$` has been removed in order to prevent build failures on
Circle CI (despite not being able to hit these issues locally).