Fixed Netlify branch detection.
Discovered the original implementation was too naive and resulted in
false positives because this repository has a clean commit history to
begin with. When used on a repository that didn't have a clean `master`
branch history, errors would appear because Netlify appears to use a
shallow clone with a detached head during the build process.

The approach taken here is similar to what was done for Travis CI where
we add the remote URL and then fetch the feature branch we want to
verify. This ensures we are *only* looking at commits on the current
feature branch.